issue #162, refactoring Twine import
