+++ www/app/config/services.yml
@@ -4,6 +4,9 @@ parameters:
 #    parameter_name: value
 
 services:
+    admin.import.twine:
+        class: AdminBundle\ImportTwine
+        arguments: ['@service_container']
 #    service_name:
 #        class: DembeloMain\Directory\ClassName
 #        arguments: ["@another_service_name", "plain_value", "%parameter_name%"]
